requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
PyJWT==2.10.0
openai-agents==0.0.14
openai==1.55.3
python-multipart==0.0.12
python-dotenv==1.0.1
